requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py#1335937
pre-commit ~= 2.19.0
pycodestyle ~= 2.8.0
requests ~= 2.28.0
setuptools ~= 60.2.0
sqlalchemy >= 1.4.37
valorant-match-history @ git+https://github.com/make-or-break/valorant-match-history#14aa8b3
valorant-utils @ git+https://github.com/make-or-break/valorant-utils#a403ca7
